requirements.txt
openai==1.35.5
appdirs
python-dateutil